requirements.txt
tempfile
python-dotenv
fastapi
uvicorn
python-multipart
pymupdf
python-dotenv
google-generativeai